requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
azure-core==1.32.0
azure-identity==1.19.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cohere==5.13.3
cryptography==44.0.0
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
eval_type_backport==0.2.0
fastavro==1.9.7
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llama-cloud==0.1.6
llama-index==0.12.3
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.3
llama-index-embeddings-adapter==0.3.0
llama-index-embeddings-openai==0.3.1
llama-index-experimental==0.5.1
llama-index-finetuning==0.3.0
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-azure-openai==0.3.0
llama-index-llms-mistralai==0.3.0
llama-index-llms-openai==0.3.2
llama-index-multi-modal-llms-openai==0.3.0
llama-index-postprocessor-cohere-rerank==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.17
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mistralai==1.2.5
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.16.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.57.0
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pillow==11.0.0
portalocker==2.10.1
propcache==0.2.1
protobuf==5.29.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
streamlit==1.40.2
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.0
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wrapt==1.17.0
yarl==1.18.3
